# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar kits de sensores y sus datos.",
        "title": "API Hexagonal Go (Sensor Kits)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.tu-soporte.com/support",
            "email": "support@tu-dominio.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080  // Cambia esto por tu host y puerto reales (puedes usar variables de entorno)",
    "basePath": "/v1             // Base path de tu API v1",
    "paths": {
...
        "/alerts/{kit_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all alerts for a given kit ID, ordered by timestamp descending.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alerts"
                ],
                "summary": "Get alerts for a specific kit",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Kit ID",
                        "name": "kit_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerts retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entities.Alert"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Kit ID provided",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve alerts",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
        "/data/airquality/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new airquality level data record for a specific kit.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airquality Data"
                ],
                "summary": "Register airquality data record",
                "parameters": [
                    {
                        "description": "airquality data record",
                        "name": "airqualityData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterAirQualityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "airquality record registered successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.AirQualityData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body, validation failed, or invalid Kit ID",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (token missing or invalid)",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error while registering airquality record",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/data/airquality/kit/{kit_id}/minutes/{minutes}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves airquality data records for a specific kit within the last N minutes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airquality Data"
                ],
                "summary": "Get recent airquality data records",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Kit ID",
                        "name": "kit_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Number of minutes ago to fetch records from",
                        "name": "minutes",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "airquality records retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entities.AirQualityData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Kit ID or minutes parameter provided",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (token missing or invalid)",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error while retrieving airquality records",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
        "/data/light/kit/{kit_id}/minutes/{minutes}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves light data records for a specific kit within the last N minutes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Light Data"
                ],
                "summary": "Get recent light data records",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Kit ID",
                        "name": "kit_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Number of minutes ago to fetch records from",
                        "name": "minutes",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Light records retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entities.LightData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Kit ID or minutes parameter provided",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (token missing or invalid)",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error while retrieving light records",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.AirQualityData": {
            "type": "object",
            "properties": {
                "air_quality_index": {
                    "description": "Air Quality Index reading",
                    "type": "integer"
                },
                "data_id": {
                    "description": "Corresponds to data_id PK",
                    "type": "integer"
                },
                "kit_id": {
                    "description": "Foreign key to kits table",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "Timestamp from DB default",
                    "type": "string"
                }
            }
        },
        "entities.Alert": {
            "type": "object",
            "properties": {
                "alert_id": {
                    "description": "Corresponds to alert_id PK",
                    "type": "integer"
                },
                "alert_type": {
                    "description": "Type of alert (e.g., \"under_min\", \"higher_max\")",
                    "type": "string"
                },
                "kit_id": {
                    "description": "Foreign key to kits table",
                    "type": "integer"
                },
                "message": {
                    "description": "Detailed message for the alert",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Timestamp from DB default",
                    "type": "string"
                }
            }
        },
...
        "entities.LightData": {
            "type": "object",
            "properties": {
                "data_id": {
                    "description": "Corresponds to data_id PK",
                    "type": "integer"
                },
                "kit_id": {
                    "description": "Foreign key to kits table",
                    "type": "integer"
                },
                "light_level": {
                    "description": "Light level reading (e.g., lux)",
                    "type": "number"
                },
                "timestamp": {
                    "description": "Timestamp from DB default",
                    "type": "string"
                }
            }
        },
...
        "request.RegisterAirQualityRequest": {
            "type": "object",
            "required": [
                "air_quality_index",
                "kit_id"
            ],
            "properties": {
                "air_quality_index": {
                    "description": "AQI must be provided and non-negative (adjust range if needed)",
                    "type": "integer",
                    "minimum": 0
                },
                "kit_id": {
                    "description": "Kit ID must be provided and positive",
                    "type": "integer"
                }
            }
        },
...
        "responses.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}